Construct a MGPM object from a vector

MGPM stays for Mixed Gaussian Phylogenetic Model. A MGPM object represents a list of all variables that are subject to change during a probabilistic inference of such a model for a given tree and trait data (see Details). This is unlike a MGPMContext object that contains the necessary data and meta-information that remains constant (e.g. model types) or changes rarely (e.g. metaI objects) during the inference.

MGPM(s, ctx)

Arguments

s
a real (double) vector representing a MGPM. Some of the entries in this vector are coerced to integers denoting nodes in the tree, regimes or model types (see example and MGPMVector).

ctx
a MGPMContext object.

Value

an object of S3 class 'MGPM'.

Details

A MGPM can be encoded as a numerical vector: $$\vec{s}=(K, n_2,...,n_{K+1}, l_2,...,l_{K+1}, r_2,...,r_{K+1}, m_1,...,m_R, v_1,...,v_P)^T$$ Use the constructor MGPMVector to create such vectors. The vector elements are described as follows:

\(K\):

number of shifts;

\((n_2,...,n_{K+1})^T\):

shift nodes - the shifts in the model occur at points within the branches leading to the shift nodes in tip-ward direction. The corresponding locations of these points are specified by \((l_2,...,l_{K+1})^T\). The nodes \((n_2,...,n_{K+1})^T\) should be ordered according to PCMTreePreorder(ctx$tree).

\((l_2,...,l_{K+1})^T\):

offset of the shift points measured as distances from the beginnings of the branches leading to shift nodes in tip-ward direction.

\((r_2,...,r_{K+1})^T\):

regime index vector. This is an integer vector with elements among \((N+1,n_2,...,n_{K+1})^T\), indicating the regime associated with each part in the tree. The regimes are named as the shift nodes, with N+1 corresponding to the part (and regime) originating at the root. The regime \(r_1 = N+1\) is always the same and therefore omitted. It is possible to have lumped regimes, that is, different parts of the tree having the same regime. This regime-lumping must obey the following rules:

neighbor parts cannot have a lumped regime. Two parts originating at nodes \(n_i\) and \(n_j\) in the tree are called neighbor parts if they are separated solely by \(n_i\) or by \(n_j\);

to resolve the conflict between the shift nodes of the different parts covered by a lumped regime, it is established that the name of a lumped regime must equal the shift-node that appears first, according to PCMTreePreorder(ctx$tree).

\((m_1,...,m_R)^T\):

model type assignment to the unique regimes. This is an integer vector with elements between 1 and M, M denoting the total number of model types possible. Each element corresponds to an element in unique(c(N+1,r_2,...,r_{K+1}))

\((v_1,...,v_P)^T\):

real numbers passed to PCMParamLoadOrStore. This is a vectorized form of the model parameters.

See also

Examples

# The PCMBase package comes with a collection of simulated objects, which we
# can use as example.

tree <- PCMBaseTestObjects$tree.ab
model <-
  PCMExtractDimensions(PCMBaseTestObjects$model_MixedGaussian_ab, dims = 1:2)
X <- PCMBaseTestObjects$traits.ab.123[1:2, ]

# Create a MGPM context
ctx <- MGPMContext(
  X, tree,
  model = MixedGaussian(
    k = 2,
    modelTypes = MGPMDefaultModelTypes(),
    mapping = structure(1:6, names = LETTERS[1:6]),
    Sigmae_x = Args_MixedGaussian_MGPMDefaultModelTypes()$Sigmae_x))

state <- MGPM(s = c(
  3,              # K: number of shifts
  52, 46, 73,     # n_2, ..., n_{K+1}: shift nodes
  0.1, 0.4, 0.3,  # l_2, ..., l_{K+1}: offsets of the shift points relative
                  # to the beginnings of shift branches in tip-ward direction.
  52, 46, 41,     # r_2,...,r_{K+1}: regime indices corresponding to the shifts.
                  # The regime of the part starting at the root node (41) is
                  # set to 41 (not included in the list). The regime for the
                  # part 73 is again 41, meaning that this regime is lumped with
                  # the root regime. So the number of regimes is R=3.
  5, 2, 2         # model type mapping for the three regimes.
  ), ctx)