Meta-information about a tree associated with a PCM

This function pre-processes the given tree and data in order to create meta-information used during likelihood calculaiton.

PCMInfo(X, tree, model, SE = matrix(0, PCMNumTraits(model),
  PCMTreeNumTips(tree)), verbose = FALSE, preorder = NULL, ...)

Arguments

X
a k x N numerical matrix with possible NA and NaN entries. Each column of X contains the measured trait values for one species (tip in tree). Missing values can be either not-available (NA) or not existing (NaN). Thse two values have are treated differently when calculating likelihoods: see PCMPresentCoordinates.

tree
a phylo object with N tips.

model
an S3 object specifying both, the model type (class, e.g. "OU") as well as the concrete model parameter values at which the likelihood is to be calculated (see also Details).

SE
a k x N matrix specifying the standard error for each measurement in X. Alternatively, a k x k x N cube specifying an upper triangular k x k Choleski factor of the variance covariance matrix for the measurement error for each node i=1, ..., N. Default: matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree)).

verbose
logical indicating if some debug-messages should printed.

preorder
an integer vector of row-indices in tree$edge matrix as returned by PCMTreePreorder. This can be given for performance speed-up when several operations needing preorder are executed on the tree. Default : NULL.

...
additional arguments used by implementing methods.

Value

a named list with the following elements:

X

k x N matrix denoting the trait data;

VE

k x k x N array denoting the measurement error variance covariance matrix for each for each tip i = 1,...,N. See the parameter SE in PCMLik.

M

total number of nodes in the tree;

N

number of tips;

k

number of traits;

RTree

number of parts on the tree (distinct elements of tree$edge.part);

RModel

number of regimes in the model (elements of attr(model, regimes));

p

number of free parameters describing the model;

r

an integer vector corresponding to tree$edge with the regime for each branch in tree;

xi

an integer vector of 0's and 1's corresponding to the rows in tree$edge indicating the presence of a jump at the corresponding branch;

pc

a logical matrix of dimension k x M denoting the present coordinates for each node;

This list is passed to PCMLik.