Parameter types

The parameter types are divided in the following categories:

Main type

These are the "ScalarParameter", "VectorParameter" and "MatrixParameter" classes. Each model parameter must have a main type.

Scope/Omission

These are the "_Global" and "_Omitted" classes. Every parameter can be global for all regimes or local for a single regime. If not specified, local scope is assumed. In some special cases a parameter (e.g. Sigmae can be omitted from a model. This is done by adding "_Omitted" to its class attribute.

Constancy (optional)

These are the "_Fixed", "_Ones", "_Identity" and "_Zeros" classes.

Transformation (optional)

These are the "_Transformable", "_CholeskyFactor" and "_Schur" classes.

Other properties (optional)

These are the "_NonNegative", "_WithNonNegativeDiagonal", "_LowerTriangular", "_AllEqual", "_ScalarDiagonal", "_Symmetric", "_UpperTriangular", "_LowerTriangularWithDiagonal" and "_UpperTriangularWithDiagonal" classes.

is.Local(o)

is.Global(o)

is.ScalarParameter(o)

is.VectorParameter(o)

is.MatrixParameter(o)

is.WithCustomVecParams(o)

is.Fixed(o)

is.Zeros(o)

is.Ones(o)

is.Identity(o)

is.AllEqual(o)

is.NonNegative(o)

is.Diagonal(o)

is.ScalarDiagonal(o)

is.Symmetric(o)

is.UpperTriangular(o)

is.UpperTriangularWithDiagonal(o)

is.WithNonNegativeDiagonal(o)

is.LowerTriangular(o)

is.LowerTriangularWithDiagonal(o)

is.Omitted(o)

is.CholeskyFactor(o)

is.Schur(o)

is.Transformable(o)

is.Transformed(o)

is.SemiPositiveDefinite(o)

Arguments

o
an object, i.e. a PCM or a parameter object.

Value

logical indicating if the object passed is from the type appearing in the function-name.

Functions

is.Local:

is.Global:

is.ScalarParameter:

is.VectorParameter:

is.MatrixParameter:

is.WithCustomVecParams:

is.Fixed:

is.Zeros:

is.Ones:

is.Identity:

is.AllEqual:

is.NonNegative:

is.Diagonal:

is.ScalarDiagonal:

is.Symmetric:

is.UpperTriangular:

is.UpperTriangularWithDiagonal:

is.WithNonNegativeDiagonal:

is.LowerTriangular:

is.LowerTriangularWithDiagonal:

is.Omitted:

is.CholeskyFactor:

is.Schur:

is.Transformable:

is.Transformed:

is.SemiPositiveDefinite: