A data.table with time, part and regime information for the nodes in a tree

A data.table with time, part and regime information for the nodes in a tree

PCMTreeDtNodes(tree)

Arguments

tree
a phylo object with node-labels and parts

Value

a data.table with a row for each node in tree and columns as follows:

startNode the starting node of each edge or NA_integer_ for the root

endNode the end node of each edge or the root id for the root

startNodeLab the character label for the startNode

endNodeLab the character label for endNode

startTime the time (distance from the root node) for the startNode or NA_real_ for the root node

endTime the time (distance from the root node) for the endNode or NA_real_ for the root node

part the part to which the edge belongs, i.e. the part of the endNode

regime the regime to which the edge belongs, i.e. the regime of the part of the endNode

Examples

PCMTreeDtNodes(PCMBaseTestObjects$tree.ab)

#>     startNode endNode startNodeLab endNodeLab startTime   endTime part regime
#>  1:        NA      41         <NA>         41        NA 0.0000000   41      a
#>  2:        41      42           41         42 0.0000000 0.6508705   41      a
#>  3:        42      43           42         43 0.6508705 0.9088872   41      a
#>  4:        43      44           43         44 0.9088872 1.3874325   41      a
#>  5:        44       1           44          1 1.3874325 2.1537432   41      a
#>  6:        44       2           44          2 1.3874325 1.4716794   41      a
#>  7:        43      45           43         45 0.9088872 1.7842086   41      a
#>  8:        45       3           45          3 1.7842086 2.1232815   41      a
#>  9:        45       4           45          4 1.7842086 2.6236489   41      a
#> 10:        42      46           42         46 0.6508705 0.9975540   41      a
#> 11:        46      47           46         47 0.9975540 1.3313289   41      a
#> 12:        47      48           47         48 1.3313289 1.8076801   41      a
#> 13:        48       5           48          5 1.8076801 2.6998785   41      a
#> 14:        48       6           48          6 1.8076801 2.6720196   41      a
#> 15:        47      49           47         49 1.3313289 1.7213184   41      a
#> 16:        49      50           49         50 1.7213184 2.4986391   41      a
#> 17:        50      51           50         51 2.4986391 3.4592571   41      a
#> 18:        51       7           51          7 3.4592571 3.8939166   41      a
#> 19:        51       8           51          8 3.4592571 4.1717718   41      a
#> 20:        50       9           50          9 2.4986391 2.8986335   41      a
#> 21:        49      10           49         10 1.7213184 2.0466706   41      a
#> 22:        46      11           46         11 0.9975540 1.7546411   41      a
#> 23:        41      52           41         52 0.0000000 0.2026923   41      a
#> 24:        52      53           52         53 0.2026923 0.9138135   41      a
#> 25:        53      54           53         54 0.9138135 1.0355054   41      a
#> 26:        54      12           54         12 1.0355054 1.2809939   41      a
#> 27:        54      55           54         55 1.0355054 1.1788098   41      a
#> 28:        55      56           55         56 1.1788098 1.4184392   41      a
#> 29:        56      13           56         13 1.4184392 1.4773736   41      a
#> 30:        56      57           56         57 1.4184392 2.0607275   41      a
#> 31:        57      14           57         14 2.0607275 2.9369967   41      a
#> 32:        57      15           57         15 2.0607275 2.8396421   41      a
#> 33:        55      58           55         58 1.1788098 1.9761186   41      a
#> 34:        58      59           58         59 1.9761186 2.4313931   41      a
#> 35:        59      60           59         60 2.4313931 2.8414771   41      a
#> 36:        60      16           60         16 2.8414771 3.6523474   41      a
#> 37:        60      17           60         17 2.8414771 3.4464104   41      a
#> 38:        59      61           59         61 2.4313931 3.0861170   41      a
#> 39:        61      62           61         62 3.0861170 3.4393143   41      a
#> 40:        62      18           62         18 3.4393143 3.7095744   41      a
#> 41:        62      19           62         19 3.4393143 4.4319983   41      a
#> 42:        61      20           61         20 3.0861170 3.7196103   41      a
#> 43:        58      63           58         63 1.9761186 2.1893267   41      a
#> 44:        63      21           63         21 2.1893267 2.3186991   41      a
#> 45:        63      64           63         64 2.1893267 2.6674448   41      a
#> 46:        64      22           64         22 2.6674448 3.5915192   41      a
#> 47:        64      23           64         23 2.6674448 3.2662057   41      a
#> 48:        53      65           53         65 0.9138135 1.8899842   41      a
#> 49:        65      66           65         66 1.8899842 2.6217767   41      a
#> 50:        66      67           66         67 2.6217767 2.9785036   41      a
#> 51:        67      68           67         68 2.9785036 3.4099773   41      a
#> 52:        68      24           68         24 3.4099773 3.5581888   41      a
#> 53:        68      25           68         25 3.4099773 3.4230549   41      a
#> 54:        67      69           67         69 2.9785036 3.6940697   41      a
#> 55:        69      26           69         26 3.6940697 3.7972539   41      a
#> 56:        69      27           69         27 3.6940697 4.1403540   41      a
#> 57:        66      28           66         28 2.6217767 3.2618777   41      a
#> 58:        65      70           65         70 1.8899842 2.8818228   41      a
#> 59:        70      29           70         29 2.8818228 3.3774164   41      a
#> 60:        70      30           70         30 2.8818228 3.3661723   41      a
#> 61:        52      71           52         71 0.2026923 0.3761346   71      b
#> 62:        71      72           71         72 0.3761346 1.1309555   71      b
#> 63:        72      73           72         73 1.1309555 1.5848510   71      b
#> 64:        73      74           73         74 1.5848510 2.0960208   71      b
#> 65:        74      31           74         31 2.0960208 2.3035659   71      b
#> 66:        74      32           74         32 2.0960208 2.3246790   71      b
#> 67:        73      75           73         75 1.5848510 2.1805630   71      b
#> 68:        75      76           75         76 2.1805630 2.7554352   71      b
#> 69:        76      33           76         33 2.7554352 2.8324996   71      b
#> 70:        76      34           76         34 2.7554352 2.7909758   71      b
#> 71:        75      77           75         77 2.1805630 2.8233585   71      b
#> 72:        77      35           77         35 2.8233585 3.7519737   71      b
#> 73:        77      36           77         36 2.8233585 3.4214509   71      b
#> 74:        72      37           72         37 1.1309555 1.6918563   71      b
#> 75:        71      78           71         78 0.3761346 0.9021623   71      b
#> 76:        78      38           78         38 0.9021623 1.8872575   71      b
#> 77:        78      79           78         79 0.9021623 1.4098041   71      b
#> 78:        79      39           79         39 1.4098041 2.0925922   71      b
#> 79:        79      40           79         40 1.4098041 2.0113454   71      b
#>     startNode endNode startNodeLab endNodeLab startTime   endTime part regime