Global options for the PCMBase package

Global options for the PCMBase package

PCMOptions()

Value

a named list with the currently set values of the following global options:

PCMBase.Value.NA NA value for the likelihood; used in GaussianPCM to return this value in case of an error occuring during likelihood calculation. By default, this is set to as.double(NA).

PCMBase.Errors.As.Warnings a logical flag indicating if errors (occuring, e.g. during likelihood calculation) should be treated as warnings and added as an attribute "error" to returne likelihood values. Default TRUE.

PCMBase.Threshold.Lambda_ij a 0-threshold for abs(Lambda_i + Lambda_j), where Lambda_i and Lambda_j are eigenvalues of the parameter matrix H of an OU or other model. Default 1e-8. See PCMPExpxMeanExp.

PCMBase.Threshold.EV A 0-threshold for the eigenvalues of the matrix V for a given branch. The V matrix is considered singular if it has eigenvalues smaller than PCMBase.Threshold.EV or when the ratio min(svdV)/max(svdV) is below PCMBase.Threshold.SV . Default is 1e-5. Treatment of branches with singular V matrix is defined by the option PCMBase.Skip.Singular.

PCMBase.Threshold.SV A 0-threshold for min(svdV)/max(svdV), where svdV is the vector of singular values of the matrix V for a given branch. The V matrix is considered singular if it has eigenvalues smaller than PCMBase.Threshold.EV or when the ratio min(svdV)/max(svdV) is below PCMBase.Threshold.SV. Default is 1e-6. Treatment of branches with singular V matrix is defined by the option PCMBase.Skip.Singular.

PCMBase.Threshold.Skip.Singular A double indicating if a branch of shorter length with singular matrix V should be skipped during likelihood calculation. Setting this option to a higher value, together with a TRUE value for the option PCMBase.Skip.Singular will result in tolerating some parameter values resulting in singular variance covariance matrix of the transition distribution. Default 1e-4.

PCMBase.Skip.Singular A logical value indicating whether branches with singular matrix V and shorter than getOption("PCMBase.Threshold.Singular.Skip") should be skipped during likelihood calculation, adding their children L,m,r values to their parent node. Default TRUE. Note, that setting this option to FALSE may cause some models to stop working, e.g. the White model. Setting this option to FALSE will also cause errors or NA likelihood values in the case of trees with very short or 0-length branches.

PCMBase.Tolerance.Symmetric A double specifying the tolerance in tests for symmetric matrices. Default 1e-8; see also isSymmetric.

PCMBase.Lmr.mode An integer code specifying the parallel likelihood calculation mode.

PCMBase.ParamValue.LowerLimitDefault lower limit value for parameters, default setting is -10.0.

PCMBase.ParamValue.LowerLimit.NonNegativeDiagonalDefault lower limit value for parameters corresponding to non-negative diagonal elements of matrices, default setting is 0.0.

PCMBase.ParamValue.UpperLimitDefault upper limit value for parameters, default setting is 10.0.

PCMBase.Transpose.Sigma_xShould upper diagonal factors for variance-covariance rate matrices be transposed, e.g. should Sigma = t(Sigma_x) Sigma_x or, rather Sigma = Sigma_x t(Sigma_x)? Note that the two variants are not equal. The default is FALSE, meaning Sigma = Sigma_x t(Sigma_x). In this case, though Sigma_x is not the actual upper Cholesky factor of Sigma, i.e. chol(Sigma) != Sigma_x. See also chol. This option applies to parameters Sigma_x, Sigmae_x and Sigmaj_x.

Examples

PCMOptions()

#> $PCMBase.Value.NA
#> [1] NA
#> 
#> $PCMBase.Errors.As.Warnings
#> [1] TRUE
#> 
#> $PCMBase.Threshold.Lambda_ij
#> [1] 1e-08
#> 
#> $PCMBase.Threshold.EV
#> [1] 1e-05
#> 
#> $PCMBase.Threshold.SV
#> [1] 1e-06
#> 
#> $PCMBase.Threshold.Skip.Singular
#> [1] 1e-04
#> 
#> $PCMBase.Skip.Singular
#> [1] TRUE
#> 
#> $PCMBase.Tolerance.Symmetric
#> [1] 1e-08
#> 
#> $PCMBase.Lmr.mode
#> [1] 11
#> 
#> $PCMBase.ParamValue.LowerLimit.NonNegativeDiagonal
#> [1] 0
#> 
#> $PCMBase.ParamValue.LowerLimit
#> [1] -10
#> 
#> $PCMBase.ParamValue.UpperLimit
#> [1] 10
#> 
#> $PCMBase.Transpose.Sigma_x
#> [1] FALSE
#>