Scatter plot of 2-dimensional data

Scatter plot of 2-dimensional data

PCMPlotTraitData2D(X, tree, labeledTips = NULL, sizeLabeledTips = 8,
  palette = PCMColorPalette(PCMTreeNumUniqueRegimes(tree),
  PCMTreeUniqueRegimes(tree)), scaleSizeWithTime = !is.ultrametric(tree))

Arguments

X
a k x N matrix

tree
a phylo object

labeledTips
a vector of tip-numbers to label (NULL by default)

sizeLabeledTips
passed geom_text to specify the size of tip-labels for the trait-points.

palette
a named vector of colors

scaleSizeWithTime
logical indicating if the size and the transparency of the points should reflect the distance from the present (points that are farther away in time with respect to the present moment, i.e. closer to the root of the tree, are displayed smaller and more transparent.). By default this is set to !is.ultrametric(tree).

Value

a ggplot object