Expected variance-covariance matrix for each couple of tips (i,j)

Expected variance-covariance matrix for each couple of tips (i,j)

PCMVar(tree, model, W0 = matrix(0, PCMNumTraits(model),
  PCMNumTraits(model)), metaI = PCMInfo(NULL, tree, model, verbose),
  internal = FALSE, verbose = FALSE)

Arguments

tree
a phylo object with N tips.

model
an S3 object specifying both, the model type (class, e.g. "OU") as well as the concrete model parameter values at which the likelihood is to be calculated (see also Details).

W0
a numeric matrix denoting the initial k x k variance covariance matrix at the root (default is the k x k zero matrix).

metaI
a list returned from a call to PCMInfo(X, tree, model), containing meta-data such as N, M and k.

internal
a logical indicating if the per-node variance-covariances matrices for the internal nodes should be returned (see Value). Default FALSE.

verbose
logical indicating if some debug-messages should printed.

Value

If internal is FALSE, a (k x N) x (k x N) matrix W, such that k x k block W[((i-1)*k)+(1:k), ((j-1)*k)+(1:k)] equals the expected covariance matrix between tips i and j. Otherwise, a list with an element 'W' as described above and a k x M matrix element 'Wii' containing the per-node variance covariance matrix for each node: The k x k block Wii[, (i-1)*k + (1:k)] represents the variance covariance matrix for node i.